Merge 2 - Diplomacy (#13)
* Saturday Night

Fixed the issue with the Quest Data no showing up the right values. Also made a little list of NPCs to work with. A followers menu is to be filled and reacted to tomorrow.

* Sunday 21/7/19 1:00

Now the Quest menu reacts to the NPC List. Next step is to make the confirm quest button work thus leading to making the means for tracking the NPCs doing quests.. numbers, race and all..