Functional Merge - EnemyAI + NPC (Neyo) + Time + Dungeon + Diplomacy (#22)
* dafaq

fa

* Sunday Night 21/7/19

I've now finished most of the UI and the quests can be retrofitted into them.
Rujul showed me a couple of classes and I realised how jugaad everything I'd coded was, lol.
Mental note: Learn how to make classes in C#. Your future depends on it.

Last but not the least I'v finished Niyati's commission and I now have all kinds of variables for her to play with. I had to change some fundamental stuff to get that done and had to go on a bit of a doozy to get there. Ya hear that Niyati?? If you're reading this I had to burn down a whole forest for you and I was actually anxious for a while about whether things would grow back!

* Monday Evening 22/07/19

Now we have a pretty decent UI.

It relates to the time now and quests can be finished with their respective rewards.

The next step is now to have it relate to Joe's quest table and Rujul's npc list. That'll take a whole day or two to do tho.