Merge 4 - NPC(Rujul) + Time + EnemyAI + Farming + Combat(Pathiri) + NPC(Neyo)+ Combat (DGD) + Loot + BaseBuild(Shreeya) + BaseBuild(Arjun) + Joe's stuff + Diplomacy
* dafaq

fa

* a

a

* The loot rewards system works now. Yay! :D

Now to finish up on the npc calling thing.

* Tuesday Night 23/7/19

Halfway through with the NPC merging.. now it's just a matter of linking it with the quest changes.
It's a deep, deep can of worms. Lots of concentration needed for this and I can't seem to stop playing The Binding of Isaac Guitar Medley on Youtube. Life is good!