Merge 5 - BB(Arjun) + Dip(Joe) + Dip(Jonty) (#43)
* dafaq

fa

* a

a

* Wednesday night (24/7/19) 98%done

Almost theeeereee.... just gotta figure out what's keeping the bonus rewards from registering. After that if... IF all goes well we're good to go.

Also, Joe's table could be very well implemented at this point too so there's that.

* Thursday Morning 25/7/19

Found some bugs in the morning. Fixed them but dunno if there are more in between the cracks. Whatever.. to hungry for that. Breakfast time.

* Thursday Morning - I think it's almost done

Forgot to make a prefab that last time, lol. But now I have and what's more.. it's even better!

* Committed again, lol.